+++ lib/ecto/adapters/foundationdb/ecto_adapter_assigns.ex
@@ -8,13 +8,16 @@ defmodule Ecto.Adapters.FoundationDB.EctoAdapterAssigns do
 
   def assign_ready(_module, repo, futures, ready_refs, options) when is_list(ready_refs) do
     Tx.transactional(options[:prefix], fn _tx ->
-      {assign_futures_rev, futures} = filter_ready(repo, futures, ready_refs, options)
+      {labeled_futures_rev, futures} = filter_ready(repo, futures, ready_refs, options)
 
-      res = repo.await(Enum.reverse(assign_futures_rev))
+      labeled_futures = Enum.reverse(labeled_futures_rev)
+      {labels, assign_futures} = Enum.unzip(labeled_futures)
+      res = repo.await(assign_futures)
+      labeled_res = Enum.zip(labels, res)
 
-      Enum.reduce(res, {[], futures}, fn
-        {new_assigns, new_future_or_nil}, {assigns, futures} ->
-          {assigns ++ new_assigns, append_new_future(futures, new_future_or_nil)}
+      Enum.reduce(labeled_res, {[], futures}, fn
+        {label, {new_assigns, new_future_or_nil}}, {assigns, futures} ->
+          {assigns ++ new_assigns, append_new_future(futures, label, new_future_or_nil)}
       end)
     end)
   end
@@ -22,51 +25,71 @@ defmodule Ecto.Adapters.FoundationDB.EctoAdapterAssigns do
   defp filter_ready(repo, futures, ready_refs, options) do
     Enum.reduce(ready_refs, {[], futures}, fn ready_ref, {acc, futures} ->
       case async_assign_ready(__MODULE__, repo, futures, ready_ref, options) do
-        {nil, futures} ->
+        {_label, nil, futures} ->
           {acc, futures}
 
-        {assign_future, futures} ->
-          {[assign_future | acc], futures}
+        {label, assign_future, futures} ->
+          {[{label, assign_future} | acc], futures}
       end
     end)
   end
 
-  defp append_new_future(futures, nil), do: futures
-  defp append_new_future(futures, future), do: [future | futures]
+  defp append_new_future(futures, _label, nil), do: futures
+  defp append_new_future(futures, _label, future) when is_list(futures), do: [future | futures]
+
+  defp append_new_future(futures, label, future) when is_map(futures) do
+    case Map.fetch(futures, label) do
+      {:ok, f} ->
+        Future.cancel(f)
+
+      _ ->
+        :ok
+    end
+
+    Map.put(futures, label, future)
+  end
 
   def async_assign_ready(_module, repo, futures, ready_ref, options)
       when is_reference(ready_ref) do
     case Future.find_ready(futures, ready_ref) do
-      {nil, futures} ->
-        {nil, futures}
+      {label, nil, futures} ->
+        {label, nil, futures}
 
-      {future, futures} ->
-        {schema, kind, watch_options, new_watch_fn} = Future.result(future)
+      {label, found_future, futures} ->
+        {schema, kind, watch_options, new_watch_fn} = Future.result(found_future)
 
-        if not Keyword.has_key?(watch_options, :label) do
+        if is_nil(label) and not Keyword.has_key?(watch_options, :label) do
           raise """
-          To use Repo.assign_ready/3, you must have previously created a watch with a label
+          To use Repo.assign_ready/3, you must either
+
+          1. Create the original watch with a label
 
           Examples:
 
               Repo.watch(struct, label: :mykey)
               SchemaMetadata.watch_collection(MySchema, label: :mykey)
+
+          2. Call Repo.assign_ready/3 with a labeled map of futures
+
+          Example:
+
+              Repo.assign_ready(%{mykey: future}, [ready_ref], watch?: true, prefix: tenant)
           """
         end
 
         case kind do
           {:pk, pk} ->
-            async_get(repo, futures, schema, pk, watch_options, options, new_watch_fn)
+            async_get(repo, label, futures, schema, pk, watch_options, options, new_watch_fn)
 
-          {SchemaMetadata, name}
-          when name in [:inserts, :deletes, :collection, :updates, :changes] ->
-            async_all(repo, futures, schema, watch_options, options, new_watch_fn)
+          {SchemaMetadata, action}
+          when action in [:inserts, :deletes, :collection, :updates, :changes] ->
+            async_all(repo, label, futures, schema, watch_options, options, new_watch_fn)
         end
     end
   end
 
-  defp async_get(repo, futures, schema, id, watch_options, options, new_watch_fn) do
-    label = watch_options[:label]
+  defp async_get(repo, label, futures, schema, id, watch_options, options, new_watch_fn) do
+    label = label || watch_options[:label]
 
     tenant = options[:prefix]
 
@@ -80,12 +103,12 @@ defmodule Ecto.Adapters.FoundationDB.EctoAdapterAssigns do
           {[{label, struct_or_nil}], new_future}
         end)
 
-      {assign_future, futures}
+      {label, assign_future, futures}
     end)
   end
 
-  defp async_all(repo, futures, schema, watch_options, options, new_watch_fn) do
-    label = watch_options[:label]
+  defp async_all(repo, label, futures, schema, watch_options, options, new_watch_fn) do
+    label = label || watch_options[:label]
     query = watch_options[:query] || schema
     tenant = options[:prefix]
 
@@ -99,7 +122,7 @@ defmodule Ecto.Adapters.FoundationDB.EctoAdapterAssigns do
           {[{label, result}], new_future}
         end)
 
-      {assign_future, futures}
+      {label, assign_future, futures}
     end)
   end